add Scale to WeightScreen
change weight value from float to int

+++ core/src/main/java/com/canonal/core/data/preferences/DefaultPreferences.kt
@@ -22,9 +22,9 @@ class DefaultPreferences(
             .apply()
     }
 
-    override fun saveWeight(weight: Float) {
+    override fun saveWeight(weight: Int) {
         sharedPref.edit()
-            .putFloat(Preferences.KEY_WEIGHT, weight)
+            .putInt(Preferences.KEY_WEIGHT, weight)
             .apply()
     }
 
@@ -67,7 +67,7 @@ class DefaultPreferences(
     override fun loadUserInfo(): UserInfo {
         val age = sharedPref.getInt(Preferences.KEY_AGE, -1)
         val height = sharedPref.getInt(Preferences.KEY_HEIGHT, -1)
-        val weight = sharedPref.getFloat(Preferences.KEY_WEIGHT, -1f)
+        val weight = sharedPref.getInt(Preferences.KEY_WEIGHT, -1)
         val genderString = sharedPref.getString(Preferences.KEY_GENDER, null)
         val activityLevelString = sharedPref.getString(Preferences.KEY_ACTIVITY_LEVEL, null)
         val goalTypeString = sharedPref.getString(Preferences.KEY_GOAL_TYPE, null)
